Adjustement in the Model, changed all ID to Long. Added Repos, Proxies for NiveauEtude and Filiere. Removed the Username from User.

+++ src/main/java/com/gsr/myschool/common/client/proxy/CandidatProxy.java
@@ -25,9 +25,9 @@ import java.util.Date;
 
 @ProxyFor(Candidat.class)
 public interface CandidatProxy extends ValueProxy {
-    Integer getId();
+    Long getId();
 
-    void setId(Integer id);
+    void setId(Long id);
 
     String getFirstname();
